code/Backend_Server/Report-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
reportlab
openpyxl